ml2_drivers: add new option to change vhost naming convention
Libvirt tries to retrieve the name of the interface of the vhost user port
with the file name of the vhost-user socket. See the link below.

The vhost socket file name is provided by the ml2 mechanism drivers to
nova. This commit adds a new option for ml2 mechanism driver to change
the naming convention of vhost socket file

Thanks to that, libvirt is able to retrieve the interface name.

Example:

 # virsh domiflist 9
 Interface  Type       Source     Model       MAC
 -------------------------------------------------------
 tap585780dc-cb bridge     br-int     virtio      fa:16:3e:e6:67:71

Link: https://github.com/libvirt/libvirt/commit/57b5e27d3d74f8dbb77ac909b405d99d555eca46

Signed-off-by: Maxime Leroy <maxime.leroy@6wind.com>
Acked-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>