ml2_drivers: fix train support
Since this commit
https://opendev.org/openstack/neutron/commit/e12580f602f81e911e1896f6cec2b26fe3d1fbf9

try_to_bind_segment_for_agent method in ml2_fp/lb_ovs mech drivers
always returns false. As a consequence accelerated port are never
created.

This can be explained because we call check_segment_for_agent with ovs
agent, and this method check if the agent type of the mech driver is the
same that agent_type of the agent provided. In this the agent_type of
ovs-fp mech driver is FP agent and the agent_type of the ovs agent is
ovs. It's why the test is failing.

To workaround the problem, let's override temporaly the self.agent_type
to ovs before calling check_segment_for_agent.

Signed-off-by: Maxime Leroy <maxime.leroy@6wind.com>
Acked-by: Christophe Grosse <christophe.grosse@6wind.com>